cadastro de usuario

+++ code/src/main.cpp
@@ -31,7 +31,7 @@ int main(int argc, char const *argv[]) {
     vector<MenuItemSet> menuItems;
 
     menuItems.push_back(MenuItemSet("Login", make_shared<LoginController>()));
-    menuItems.push_back(MenuItemSet("Cadastro", make_shared<UserController>(), (int)ControllerActionEnum::CREATE));
+    menuItems.push_back(MenuItemSet("Cadastro", make_shared<UserController>(), ControllerActionEnum::CREATE));
     menuItems.push_back(MenuItemSet("Sair", nullptr));
     
     MenuController menuController("Menu Principal", menuItems);
